+++ src/icp_basic_dex_frontend/src/App.jsx
@@ -5,12 +5,10 @@ import { Actor, HttpAgent } from "@dfinity/agent";
 import { AuthClient } from "@dfinity/auth-client";
 import { Principal } from '@dfinity/principal';
 
-import { canisterId as IICanisterID }
-  from "../../declarations/internet_identity";
 import { canisterId as DEXCanisterId }
-  from "../../declarations/icp_basic_dex_backend";
+  from '../../declarations/icp_basic_dex_backend';
 import { idlFactory as DEXidlFactory }
-  from "../../declarations/icp_basic_dex_backend/icp_basic_dex_backend.did.js";
+  from '../../declarations/icp_basic_dex_backend/icp_basic_dex_backend.did.js';
 import { canisterId as HogeDIP20canisterId }
   from '../../declarations/HogeDIP20';
 import { idlFactory as HogeidlFactory }
@@ -25,18 +23,18 @@ import { UserBoard } from './components/UserBoard';
 import { PlaceOrder } from './components/PlaceOrder';
 import { ListOrder } from './components/ListOrder';
 
-// import { UserTokensProvider } from './context/UserTokens';
-
 const App = () => {
 
   const tokenCanisters = [
     {
       canisterName: 'HogeDIP20',
+      tokenSymbol: 'THG',
       factory: HogeidlFactory,
       canisterId: HogeDIP20canisterId,
     },
     {
       canisterName: 'PiyoDIP20',
+      tokenSymbol: 'TPY',
       factory: PiyoidlFactory,
       canisterId: PiyoDIP20canisterId,
     },
@@ -50,136 +48,7 @@ const App = () => {
 
   const [orderList, setOrderList] = useState([]);
 
-  const [order, setOrder] = useState({
-    from: '',
-    fromAmount: 0,
-    to: '',
-    toAmount: 0,
-  })
-
-  const handleChangeOrder = (event) => {
-    setOrder((prevState) => {
-      return {
-        ...prevState,
-        [event.target.name]: event.target.value,
-      };
-    });
-  };
-
-  const handleSubmitOrder = async (event) => {
-    event.preventDefault();
-    console.log(order);
-
-    const DEXActor = Actor.createActor(DEXidlFactory, {
-      agent,
-      canisterId: DEXCanisterId,
-    });
-
-    let fromPrincipal;
-    if (order.from === "THG") {
-      fromPrincipal = Principal.fromText(HogeDIP20canisterId);
-    } else {
-      fromPrincipal = Principal.fromText(PiyoDIP20canisterId);
-    }
-
-    let toPrincipal;
-    if (order.to === "THG") {
-      toPrincipal = Principal.fromText(HogeDIP20canisterId);
-    } else {
-      toPrincipal = Principal.fromText(PiyoDIP20canisterId);
-    }
-
-    try {
-      const resultPlace
-        = await DEXActor.placeOrder(
-          fromPrincipal,
-          Number(order.fromAmount),
-          toPrincipal,
-          Number(order.toAmount),
-        );
-      // Check Error
-      if (!resultPlace.Ok) {
-        alert(`Error: ${Object.keys(resultPlace.Err)[0]}`);
-        return;
-      }
-
-      // Update Order List
-      const updateOrders = await DEXActor.getOrders();
-      setOrderList(updateOrders);
-
-      console.log(`Created order:  ${resultPlace.Ok[0].id}`);
-    } catch (error) {
-      console.log(`handleSubmitOrder: ${error} `);
-    }
-  };
-
-  // Buy order hander
-  const handleBuyOrder = async (order) => {
-    // Create DEX actor
-    const DEXActor = Actor.createActor(DEXidlFactory, {
-      agent,
-      canisterId: DEXCanisterId,
-    });
-    try {
-
-      // Call placeOrder
-      const resultPlace
-        = await DEXActor.placeOrder(
-          order.to,
-          Number(order.toAmount),
-          order.from,
-          Number(order.fromAmount),
-        );
-
-      // Check Error
-      if (!resultPlace.Ok) {
-        alert(`Error: ${Object.keys(resultPlace.Err)[0]}`);
-        return;
-      }
-
-      // Update order list
-      const updateOrders = await DEXActor.getOrders();
-      setOrderList(updateOrders);
-
-      // Update user balances
-      const tokens = await getUserTokens(agent, Principal.fromText(currentPrincipalId));
-      setUserTokens(tokens);
-
-      console.log("Trade Successful!");
-    } catch (error) {
-      console.log(`handleBuyOrder: ${error} `);
-    };
-  };
-
-
-  // Cancel order handler
-  const handleCancelOrder = async (id) => {
-    try {
-      const DEXActor = Actor.createActor(DEXidlFactory, {
-        agent,
-        canisterId: DEXCanisterId,
-      });
-
-      // Call cancelOrder
-      const resultCancel = await DEXActor.cancelOrder(id);
-
-      // Check Error
-      if (!resultCancel.Ok) {
-        alert(`Error: ${Object.keys(resultCancel.Err)}`);
-        return;
-      }
-
-      // Update orderbook
-      const updateOrders = await DEXActor.getOrders();
-      setOrderList(updateOrders);
-
-      console.log(`Canceled order ID: ${resultCancel.Ok}`);
-    } catch (error) {
-      console.log(`handleCancelOrder: ${error}`);
-    }
-  }
-
-  const getUserTokens = async (agent, principal) => {
+  const updateUserTokens = async (agent, principal) => {
     const DEXActor = Actor.createActor(DEXidlFactory, {
       agent,
       canisterId: DEXCanisterId,
@@ -212,11 +81,11 @@ const App = () => {
       }
       tokens.push(userToken);
     }
-    // setUserTokens(tokens);
-    return tokens;
+    // return tokens;
+    setUserTokens(tokens);
   }
 
-  const getOrders = async (agent) => {
+  const updateOrderList = async (agent) => {
     const DEXActor = Actor.createActor(DEXidlFactory, {
       agent,
       canisterId: DEXCanisterId,
@@ -224,62 +93,20 @@ const App = () => {
 
     // Set Order list
     const orders = await DEXActor.getOrders();
-    setOrderList(orders);
+    const createdOrderList = orders.map((order) => {
+      return {
+        id: order.id,
+        from: order.from,
+        fromSymbol: tokenCanisters.find(e => e.canisterId === order.from.toString()).tokenSymbol,
+        fromAmount: order.fromAmount,
+        to: order.to,
+        toSymbol: tokenCanisters.find(e => e.canisterId === order.to.toString()).tokenSymbol,
+        toAmount: order.toAmount,
+      }
+    })
+    // return orders;
+    setOrderList(createdOrderList);
   }
-
-  // Login Internet Identity handler
-  const handleLogin = async () => {
-    // Autofills the <input> for the II Url to point to the correct canister.
-    console.log(`process.env.DFX_NETWORK: ${process.env.DFX_NETWORK}`);
-
-    let iiUrl;
-    if (process.env.DFX_NETWORK === "local") {
-      iiUrl = `http://localhost:8000/?canisterId=${IICanisterID}`;
-    } else if (process.env.DFX_NETWORK === "ic") {
-      iiUrl = `https://${IICanisterID}.ic0.app`;
-    } else {
-      iiUrl = `https://${IICanisterID}.dfinity.network`;
-    }
-    // Start Login process.
-    // First we have to create and AuthClient.
-    const authClient = await AuthClient.create();
-
-    // Login with Internet Identity.
-    await new Promise((resolve, reject) => {
-      authClient.login({
-        identityProvider: iiUrl,
-        onSuccess: resolve,
-        onError: reject,
-      });
-    });
-
-
-    // Get the identity from the auth client:
-    const identity = authClient.getIdentity();
-    // Using the identity obtained from the auth client,
-    // we can create an agent to interact with the IC.
-    const newAgent = new HttpAgent({ identity });
-
-    if (process.env.DFX_NETWORK === "local") {
-      newAgent.fetchRootKey();
-    }
-    setAgent(newAgent);
-    // Using the interface description of our webapp,
-    // we create an Actor that we use to call the service methods.
-    const DEXActor = Actor.createActor(DEXidlFactory, {
-      newAgent,
-      canisterId: DEXCanisterId,
-    });
-    const principal = await authClient.getIdentity().getPrincipal();
-
-    // Get information about the tokens held by the Logged-in user.
-    const tokens = await getUserTokens(newAgent, principal);
-    setUserTokens(tokens);
-    // Set Order list
-    getOrders(newAgent);
-
-    setCurrentPrincipalId(principal.toText());
-  };
 
   const checkClientIdentity = async () => {
     try {
@@ -299,9 +126,8 @@ const App = () => {
           newAgent.fetchRootKey();
         }
 
-        const tokens = await getUserTokens(newAgent, principal);
-        setUserTokens(tokens);
-        getOrders(newAgent);
+        updateUserTokens(newAgent, principal);
+        updateOrderList(newAgent);
         setAgent(newAgent);
       } else {
         console.log(`isAuthenticated: ${resultAuthenticated}`);
@@ -310,88 +136,6 @@ const App = () => {
       console.log(`checkClientIdentity: ${error}`);
     }
   }
-
-  const handleDeposit = async (updateIndex) => {
-    // const [userTokens, setUserTokens] = useUserTokensContext();
-    const tokenActor = Actor.createActor(tokenCanisters[updateIndex].factory, {
-      agent,
-      canisterId: tokenCanisters[updateIndex].canisterId,
-    });
-
-    const DEXActor = Actor.createActor(DEXidlFactory, {
-      agent,
-      canisterId: DEXCanisterId,
-    });
-
-    try {
-      // Approve user token transfer by DEX.
-      const resultApprove
-        = await tokenActor.approve(Principal.fromText(DEXCanisterId), 5000);
-      console.log(`resultApprove: ${resultApprove.Ok}`);
-      // Deposit token from token canister to DEX.
-      const resultDeposit
-        = await DEXActor.deposit(Principal.fromText(tokenCanisters[updateIndex].canisterId));
-      console.log(`resultDeposit: ${resultDeposit.Ok}`);
-      // Get updated balance of token Canister.
-      const balance
-        = await tokenActor.balanceOf(Principal.fromText(currentPrincipalId));
-      // Get updated balance in DEX.
-      const dexBalance
-        = await DEXActor.getBalance(Principal.fromText(tokenCanisters[updateIndex].canisterId));
-
-      // Set new user infomation.
-      setUserTokens(
-        userTokens.map((userToken, index) => (
-          index === updateIndex ? {
-            symbol: userToken.symbol,
-            balance: balance.toString(),
-            dexBalance: dexBalance.toString(),
-            fee: userToken.fee,
-          } : userToken))
-      );
-
-    } catch (error) {
-      console.log(`handleDeposit: ${error} `);
-    }
-  };
-
-  const handleWithdraw = async (updateIndex) => {
-    const tokenActor = Actor.createActor(tokenCanisters[updateIndex].factory, {
-      agent,
-      canisterId: tokenCanisters[updateIndex].canisterId,
-    });
-
-    const DEXActor = Actor.createActor(DEXidlFactory, {
-      agent,
-      canisterId: DEXCanisterId,
-    });
-
-    try {
-      const resultWithdraw
-        = await DEXActor.withdraw(Principal.fromText(tokenCanisters[updateIndex].canisterId), 5000);
-      console.log(`resultWithdraw: ${resultWithdraw.Ok}`);
-      // Get updated balance of token Canister.
-      const balance
-        = await tokenActor.balanceOf(Principal.fromText(currentPrincipalId));
-      // Get updated balance in DEX.
-      const dexBalance
-        = await DEXActor.getBalance(Principal.fromText(tokenCanisters[updateIndex].canisterId));
-
-      // Set new user infomation.
-      setUserTokens(
-        userTokens.map((userToken, index) => (
-          index === updateIndex ? {
-            symbol: userToken.symbol,
-            balance: balance.toString(),
-            dexBalance: dexBalance.toString(),
-            fee: userToken.fee,
-          } : userToken))
-      );
-
-    } catch (error) {
-      console.log(`handleWithdraw: ${error} `);
-    }
-  };
 
   // ページがリロードされた時、以下の関数を実行
   useEffect(() => {
@@ -402,28 +146,29 @@ const App = () => {
   return (
     <>
       <Header
-        handleLogin={handleLogin}
+        updateOrderList={updateOrderList}
+        updateUserTokens={updateUserTokens}
+        setAgent={setAgent}
+        setCurrentPrincipalId={setCurrentPrincipalId}
       />
       <main className="app">
-        {/* <UserTokensProvider> */}
         <UserBoard
-          Actor={Actor}
+          agent={agent}
           tokenCanisters={tokenCanisters}
           currentPrincipalId={currentPrincipalId}
           userTokens={userTokens}
-          handleDeposit={handleDeposit}
-          handleWithdraw={handleWithdraw}
+          setUserTokens={setUserTokens}
         />
-        {/* </UserTokensProvider> */}
         <PlaceOrder
+          agent={agent}
+          tokenCanisters={tokenCanisters}
           currentPrincipalId={currentPrincipalId}
-          handleChangeOrder={handleChangeOrder}
-          handleSubmitOrder={handleSubmitOrder}
+          updateOrderList={updateOrderList}
         />
         <ListOrder
+          agent={agent}
           orderList={orderList}
-          handleBuyOrder={handleBuyOrder}
-          handleCancelOrder={handleCancelOrder}
+          updateOrderList={updateOrderList}
         />
       </main>
     </>
