+++ src/icp_basic_dex_frontend/src/components/ListOrder.jsx
@@ -1,7 +1,78 @@
 import React from 'react';
 
+import { Actor } from '@dfinity/agent';
+import { canisterId as DEXCanisterId }
+  from '../../../declarations/icp_basic_dex_backend';
+import { idlFactory as DEXidlFactory }
+  from '../../../declarations/icp_basic_dex_backend/icp_basic_dex_backend.did.js';
+import { Principal } from '@dfinity/principal';
+
 export const ListOrder = (props) => {
-  const { orderList, handleBuyOrder, handleCancelOrder } = props;
+  const { agent, orderList, updateOrderList } = props;
+
+  // Buy order hander
+  const handleBuyOrder = async (order) => {
+    // Create DEX actor
+    const DEXActor = Actor.createActor(DEXidlFactory, {
+      agent,
+      canisterId: DEXCanisterId,
+    });
+    try {
+
+      // Call placeOrder
+      const resultPlace
+        = await DEXActor.placeOrder(
+          order.to,
+          Number(order.toAmount),
+          order.from,
+          Number(order.fromAmount),
+        );
+
+      // Check Error
+      if (!resultPlace.Ok) {
+        alert(`Error: ${Object.keys(resultPlace.Err)[0]}`);
+        return;
+      }
+
+      // Update order list
+      updateOrderList(agent);
+
+      // Update user balances
+      updateUserTokens(agent, Principal.fromText(currentPrincipalId));
+
+      console.log("Trade Successful!");
+    } catch (error) {
+      console.log(`handleBuyOrder: ${error} `);
+    };
+  };
+
+
+  // Cancel order handler
+  const handleCancelOrder = async (id) => {
+    try {
+      const DEXActor = Actor.createActor(DEXidlFactory, {
+        agent,
+        canisterId: DEXCanisterId,
+      });
+
+      // Call cancelOrder
+      const resultCancel = await DEXActor.cancelOrder(id);
+
+      // Check Error
+      if (!resultCancel.Ok) {
+        alert(`Error: ${Object.keys(resultCancel.Err)}`);
+        return;
+      }
+
+      // Update orderbook
+      updateOrderList(agent);
+
+      console.log(`Canceled order ID: ${resultCancel.Ok}`);
+    } catch (error) {
+      console.log(`handleCancelOrder: ${error}`);
+    }
+  }
+
 
   return (
     <div className="order-list" style={{ backgroundColor: "rgb(8, 2, 38)" }}>
@@ -19,10 +90,10 @@ export const ListOrder = (props) => {
           {orderList.map((order, index) => {
             return (
               <tr key={`${index}: ${order.token} `} >
-                <td data-th="From">{order.from.toString()}</td>
+                <td data-th="From">{order.fromSymbol}</td>
                 <td data-th="Amount">{order.fromAmount.toString()}</td>
                 <td>→</td>
-                <td data-th="To">{order.to.toString()}</td>
+                <td data-th="To">{order.toSymbol}</td>
                 <td data-th="Amount">{order.toAmount.toString()}</td>
                 <td data-th="Action">
                   <div>
