+++ src/icp_basic_dex_frontend/src/components/PlaceOrder.jsx
@@ -1,7 +1,73 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { Principal } from '@dfinity/principal';
+import { Actor } from "@dfinity/agent";
+import { canisterId as DEXCanisterId }
+  from '../../../declarations/icp_basic_dex_backend';
+import { idlFactory as DEXidlFactory }
+  from '../../../declarations/icp_basic_dex_backend/icp_basic_dex_backend.did.js';
 
 export const PlaceOrder = (props) => {
-  const { currentPrincipalId, handleChangeOrder, handleSubmitOrder } = props;
+  const {
+    agent,
+    currentPrincipalId,
+    tokenCanisters,
+    updateOrderList,
+  } = props;
+
+  const [order, setOrder] = useState({
+    from: '',
+    fromAmount: 0,
+    to: '',
+    toAmount: 0,
+  })
+
+  const handleChangeOrder = (event) => {
+    setOrder((prevState) => {
+      return {
+        ...prevState,
+        [event.target.name]: event.target.value,
+      };
+    });
+  };
+
+  const handleSubmitOrder = async (event) => {
+    event.preventDefault();
+    console.log(order);
+
+    const DEXActor = Actor.createActor(DEXidlFactory, {
+      agent,
+      canisterId: DEXCanisterId,
+    });
+
+    // Get from token canister
+    const fromTokenCanister = tokenCanisters.find(e => e.tokenSymbol === order.from);
+    const fromPrincipal = fromTokenCanister.canisterId;
+    // Get to token canister
+    const toTokenCanister = tokenCanisters.find(e => e.tokenSymbol === order.to);
+    const toPrincipal = toTokenCanister.canisterId;
+
+    try {
+      const resultPlace
+        = await DEXActor.placeOrder(
+          Principal.fromText(fromPrincipal),
+          Number(order.fromAmount),
+          Principal.fromText(toPrincipal),
+          Number(order.toAmount),
+        );
+      // Check Error
+      if (!resultPlace.Ok) {
+        alert(`Error: ${Object.keys(resultPlace.Err)[0]}`);
+        return;
+      }
+
+      // Update Order List
+      updateOrderList(agent);
+
+      console.log(`Created order:  ${resultPlace.Ok[0].id}`);
+    } catch (error) {
+      console.log(`handleSubmitOrder: ${error} `);
+    }
+  };
 
   return (
     <>
@@ -9,7 +75,6 @@ export const PlaceOrder = (props) => {
         <div className="create-order-area">
           <div className="title">
             <p>PLACE ORDER</p>
-            {/* <button>+</button> */}
           </div>
           <form className="form" onSubmit={handleSubmitOrder} >
             <div>
